+++ src/targets/index.ts
@@ -13,6 +13,7 @@ import { PypiTarget } from './pypi';
 import { RegistryTarget } from './registry';
 import { AwsLambdaLayerTarget } from './awsLambdaLayer';
 import { UpmTarget } from './upm';
+import { MavenTarget } from './maven';
 import { SymbolCollector } from './symbolCollector';
 
 export const TARGET_MAP: { [key: string]: typeof BaseTarget } = {
@@ -30,6 +31,7 @@ export const TARGET_MAP: { [key: string]: typeof BaseTarget } = {
   registry: RegistryTarget,
   'aws-lambda-layer': AwsLambdaLayerTarget,
   upm: UpmTarget,
+  maven: MavenTarget,
   'symbol-collector': SymbolCollector,
 };